Split SvgMerge
