+++ Source/Filter Effects/feMerge/SvgMerge.Drawing.cs
@@ -1,0 +1,27 @@
+#if !NO_SDC
+using System.Drawing;
+using System.Linq;
+
+namespace Svg.FilterEffects
+{
+    public partial class SvgMerge : SvgFilterPrimitive
+    {
+        public override void Process(ImageBuffer buffer)
+        {
+            var children = this.Children.OfType<SvgMergeNode>().ToList();
+            var inputImage = buffer[children.First().Input];
+            var result = new Bitmap(inputImage.Width, inputImage.Height);
+            using (var g = Graphics.FromImage(result))
+            {
+                foreach (var child in children)
+                {
+                    g.DrawImage(buffer[child.Input], new Rectangle(0, 0, inputImage.Width, inputImage.Height),
+                        0, 0, inputImage.Width, inputImage.Height, GraphicsUnit.Pixel);
+                }
+                g.Flush();
+            }
+            buffer[this.Result] = result;
+        }
+    }
+}
+#endif
